Streamlit/requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
folium==0.18.0
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lightgbm==4.5.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mysql-connector-python==8.0.33
narwhals==1.11.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
protobuf==3.20.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
scikit-learn==1.5.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
streamlit-folium==0.23.1
tenacity==9.0.0
tensorflow==2.17.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
xyzservices==2024.9.0